Move com.typesafe.sbtrc.client to sbt.client.impl
This way it can use private[sbt] stuff and itself be private[sbt]
(but privatization not done yet, just moving)
