+++ client/src/main/scala/sbt/client/impl/DebugClient.scala
@@ -1,5 +1,5 @@
-package com.typesafe.sbtrc
-package client
+
+package sbt.client.impl
 
 import sbt.impl.ipc
 import java.net._
