+++ client/src/main/scala/sbt/client/impl/SimpleChannel.scala
@@ -1,5 +1,5 @@
-package com.typesafe.sbtrc
-package client
+
+package sbt.client.impl
 
 import sbt.impl.ipc
 import sbt.protocol
